use pico_static_assert to replace static_assert
pico's static_assert totally replies on STD library. However, in the
github CI environment, it doesn't work. The reason is on the local
building environment, it is support C++11 but github CI doesn't have it.
The 'static_assert' only can be used if the C++ >= 11.
So we define a pico_static_assert to replace the direct call from
static_assert, and if the C++ is >= 11, we call static_assert; if not,
we will only provide an empty pico_static_assert macro to do nothing but
only satisfy with the gramma.